+++ read-stream.js
