+++ write-stream.js
